Relocate files base on new import method
